Complete gameplay
The user can now  play the whole game. With Hi Score display.

Currently, if the selected random cards are all choices that have been
made before. It will remove one and add an unchosen selection. However,
it always adds it at the end of the array. Need to add an array shuffle
function.