Shuffle array after unselected mountain pushed onto it
When the randomly selected mountains are all mountains the player has
selected before. I simply pop one off the end, find a new unselected
mountain and push onto the end. The result is that the unselected mountain
is always at the end. So now I have a hacky kind of array shuffle. It simply
sorts the array, and the sort function returns a random positive or
negative integer.